moved employee class to user class
